Fixed masked support with cutoff

+++ XR_APILAYER_NOVENDOR_steamvr_passthrough/config_manager.cpp
@@ -138,7 +138,7 @@ void ConfigManager::ParseConfig_Stereo()
 	m_configCustomStereo.StereoCutoutFactor = (float)m_iniData.GetDoubleValue("StereoCustom", "StereoCutoutFactor", m_configCustomStereo.StereoCutoutFactor);
 	m_configCustomStereo.StereoCutoutOffset = (float)m_iniData.GetDoubleValue("StereoCustom", "StereoCutoutOffset", m_configCustomStereo.StereoCutoutOffset);
 	m_configCustomStereo.StereoDisparityFilterWidth = (int)m_iniData.GetLongValue("StereoCustom", "StereoDisparityFilterWidth", m_configCustomStereo.StereoDisparityFilterWidth);
-	m_configCustomStereo.StereoCutoutFilterWidth = (int)m_iniData.GetLongValue("StereoCustom", "StereoCutoutFilterWidth", m_configCustomStereo.StereoCutoutFilterWidth);
+	m_configCustomStereo.StereoCutoutFilterWidth = (float)m_iniData.GetDoubleValue("StereoCustom", "StereoCutoutFilterWidth", m_configCustomStereo.StereoCutoutFilterWidth);
 
 	m_configCustomStereo.StereoBlockSize = m_iniData.GetLongValue("StereoCustom", "StereoBlockSize", m_configCustomStereo.StereoBlockSize);
 	m_configCustomStereo.StereoMinDisparity = m_iniData.GetLongValue("StereoCustom", "StereoMinDisparity", m_configCustomStereo.StereoMinDisparity);
@@ -226,7 +226,7 @@ void ConfigManager::UpdateConfig_Stereo()
 	m_iniData.SetDoubleValue("StereoCustom", "StereoCutoutFactor", m_configCustomStereo.StereoCutoutFactor);
 	m_iniData.SetDoubleValue("StereoCustom", "StereoCutoutOffset", m_configCustomStereo.StereoCutoutOffset);
 	m_iniData.SetLongValue("StereoCustom", "StereoDisparityFilterWidth", m_configCustomStereo.StereoDisparityFilterWidth);
-	m_iniData.SetLongValue("StereoCustom", "StereoCutoutFilterWidth", m_configCustomStereo.StereoCutoutFilterWidth);
+	m_iniData.SetDoubleValue("StereoCustom", "StereoCutoutFilterWidth", m_configCustomStereo.StereoCutoutFilterWidth);
 
 	m_iniData.SetLongValue("StereoCustom", "StereoBlockSize", m_configCustomStereo.StereoBlockSize);
 	m_iniData.SetLongValue("StereoCustom", "StereoMinDisparity", m_configCustomStereo.StereoMinDisparity);
